+++ pax-compiler/src/helpers.rs
@@ -1,8 +1,24 @@
+use serde::Deserialize;
 use std::fs;
-use std::path::Path;
+use std::path::{Path, PathBuf};
+use std::process::Command;
 use std::str::FromStr;
+use std::sync::{Arc, Mutex};
 use toml_edit;
 use toml_edit::Document;
+
+use crate::ALL_PKGS;
+
+#[derive(Debug, Deserialize)]
+struct Metadata {
+    packages: Vec<Package>,
+}
+
+#[derive(Debug, Deserialize)]
+struct Package {
+    name: String,
+    version: String,
+}
 
 pub fn set_path_on_pax_dependencies(full_path: &Path) {
     // Read the Cargo.toml
@@ -124,5 +140,132 @@ pub fn update_pax_dependency_versions(doc: &mut Document, ctx_version: &str) {
                 *dep_entry = toml_edit::Item::from_str(&dep_string).unwrap_or_default();
             }
         }
+    }
+}
+
+const ERR_LOCK: &str = "Failed to lock process_child_ids mutex";
+
+pub fn wait_with_output(
+    process_child_ids: &Arc<Mutex<Vec<u64>>>,
+    child: std::process::Child,
+) -> std::process::Output {
+    let child_id: u64 = child.id().into();
+
+    // Push the child_id to the shared process_child_ids vector
+    process_child_ids.lock().expect(ERR_LOCK).push(child_id);
+
+    // Wait for the child process to complete
+    let output = child
+        .wait_with_output()
+        .expect("Failed to wait for child process");
+
+    // Ensure the child ID is removed after completion
+    process_child_ids
+        .lock()
+        .expect(ERR_LOCK)
+        .retain(|&id| id != child_id);
+
+    output
+}
+
+pub fn get_or_create_pax_directory(working_dir: &str) -> PathBuf {
+    let working_path = std::path::Path::new(working_dir).join(".pax");
+    std::fs::create_dir_all(&working_path).unwrap();
+    fs::canonicalize(working_path).unwrap()
+}
+
+pub fn get_version_of_whitelisted_packages(path: &str) -> Result<String, &'static str> {
+    let mut cmd = Command::new("cargo");
+    let output = cmd
+        .arg("metadata")
+        .arg("--format-version=1")
+        .current_dir(path)
+        .output()
+        .expect("Failed to execute `cargo metadata`");
+
+    if !output.status.success() {
+        eprintln!("{}", String::from_utf8_lossy(&output.stderr));
+        panic!("Failed to get metadata from Cargo");
+    }
+
+    let metadata: Metadata =
+        serde_json::from_slice(&output.stdout).expect("Failed to parse JSON from `cargo metadata`");
+
+    let mut tracked_version: Option<String> = None;
+
+    for package in &metadata.packages {
+        if ALL_PKGS.contains(&package.name.as_str()) {
+            if let Some(ref version) = tracked_version {
+                if package.version != *version {
+                    panic!(
+                        "Version mismatch for {}: expected {}, found {}",
+                        package.name, version, package.version
+                    );
+                }
+            } else {
+                tracked_version = Some(package.version.clone());
+            }
+        }
+    }
+
+    tracked_version.ok_or("Cannot build a Pax project without a `pax-*` dependency somewhere in your project's dependency graph.  Add e.g. `pax-lang` to your Cargo.toml to resolve this error.")
+}
+
+/// Helper recursive fs copy method, like fs::copy, but suited for our purposes.
+/// Includes ability to ignore directories by name during recursion.
+pub fn copy_dir_recursively(
+    src: &Path,
+    dest: &Path,
+    ignore_list: &[&str],
+) -> Result<(), Box<dyn std::error::Error>> {
+    if src.is_dir() {
+        // If the directory name is in the ignore list, we skip this directory
+        if ignore_list.contains(&src.file_name().unwrap().to_str().unwrap()) {
+            return Ok(());
+        }
+
+        // Create the corresponding directory in the destination,
+        // and copy its contents recursively
+        fs::create_dir_all(dest)?;
+        for entry in fs::read_dir(src)? {
+            let entry = entry?;
+            let path = entry.path();
+            let dest_child = dest.join(path.file_name().ok_or("Invalid file name")?);
+            copy_dir_recursively(&path, &dest_child, ignore_list)?;
+        }
+    } else {
+        // If source is a file, just copy it to the destination
+        fs::copy(src, dest)?;
+    }
+    Ok(())
+}
+
+/// Pulled from host Cargo.toml
+pub struct HostCrateInfo {
+    /// for example: `pax-example`
+    pub name: String,
+    /// for example: `pax_example`
+    pub identifier: String,
+    /// for example: `some_crate::pax_reexports`,
+    pub import_prefix: String,
+}
+
+pub fn get_host_crate_info(cargo_toml_path: &Path) -> HostCrateInfo {
+    let existing_cargo_toml = toml_edit::Document::from_str(
+        &fs::read_to_string(fs::canonicalize(cargo_toml_path).unwrap()).unwrap(),
+    )
+    .expect("Error loading host Cargo.toml");
+
+    let name = existing_cargo_toml["package"]["name"]
+        .as_str()
+        .unwrap()
+        .to_string();
+    let identifier = name.replace("-", "_"); //NOTE: perhaps this could be less naive?
+    let import_prefix = format!("{}::pax_reexports::", &identifier);
+
+    HostCrateInfo {
+        name,
+        identifier,
+        import_prefix,
     }
 }
