+++ pax-compiler/src/reexports/namespace_trie_node.rs
@@ -1,0 +1,129 @@
+use std::cmp::Ordering;
+use std::collections::HashMap;
+
+use itertools::Itertools;
+pub struct NamespaceTrieNode {
+    pub node_string: Option<String>,
+    pub children: HashMap<String, NamespaceTrieNode>,
+}
+
+impl PartialEq for NamespaceTrieNode {
+    fn eq(&self, other: &Self) -> bool {
+        self.node_string == other.node_string
+    }
+}
+
+impl Eq for NamespaceTrieNode {}
+
+impl PartialOrd for NamespaceTrieNode {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for NamespaceTrieNode {
+    fn cmp(&self, other: &Self) -> Ordering {
+        match (&self.node_string, &other.node_string) {
+            (Some(a), Some(b)) => a.cmp(b),
+            (Some(_), None) => Ordering::Greater,
+            (None, Some(_)) => Ordering::Less,
+            (None, None) => Ordering::Equal,
+        }
+    }
+}
+
+impl NamespaceTrieNode {
+    pub fn insert(&mut self, namespace_string: &str) {
+        let mut segments = namespace_string.split("::");
+        let first_segment = segments.next().unwrap();
+
+        let mut current_node = self;
+        current_node = current_node.get_or_create_child(first_segment);
+
+        for segment in segments {
+            current_node = current_node.get_or_create_child(segment);
+        }
+    }
+
+    pub fn get_or_create_child(&mut self, segment: &str) -> &mut NamespaceTrieNode {
+        self.children
+            .entry(segment.to_string())
+            .or_insert_with(|| NamespaceTrieNode {
+                node_string: Some(if let Some(ns) = self.node_string.as_ref() {
+                    ns.to_string() + "::" + segment
+                } else {
+                    segment.to_string()
+                }),
+                children: HashMap::new(),
+            })
+    }
+
+    pub fn serialize_to_reexports(&self) -> String {
+        "pub mod pax_reexports {\n".to_string() + &self.recurse_serialize_to_reexports(1) + "\n}"
+    }
+
+    pub fn recurse_serialize_to_reexports(&self, indent: usize) -> String {
+        let indent_str = "    ".repeat(indent);
+
+        let mut accum: String = "".into();
+
+        self.children.iter().sorted().for_each(|child| {
+            if child.1.node_string.as_ref().unwrap() == "crate" {
+                //handle crate subtrie by skipping the crate NamespaceTrieNode, traversing directly into its children
+                child.1.children.iter().sorted().for_each(|child| {
+                    if child.1.children.len() == 0 {
+                        //leaf node:  write `pub use ...` entry
+                        accum += &format!(
+                            "{}pub use {};\n",
+                            indent_str,
+                            child.1.node_string.as_ref().unwrap()
+                        );
+                    } else {
+                        //non-leaf node:  write `pub mod ...` block
+                        accum += &format!(
+                            "{}pub mod {} {{\n",
+                            indent_str,
+                            child
+                                .1
+                                .node_string
+                                .as_ref()
+                                .unwrap()
+                                .split("::")
+                                .last()
+                                .unwrap()
+                        );
+                        accum += &child.1.recurse_serialize_to_reexports(indent + 1);
+                        accum += &format!("{}}}\n", indent_str);
+                    }
+                })
+            } else {
+                if child.1.children.len() == 0 {
+                    //leaf node:  write `pub use ...` entry
+                    accum += &format!(
+                        "{}pub use {};\n",
+                        indent_str,
+                        child.1.node_string.as_ref().unwrap()
+                    );
+                } else {
+                    //non-leaf node:  write `pub mod ...` block
+                    accum += &format!(
+                        "{}pub mod {}{{\n",
+                        indent_str,
+                        child
+                            .1
+                            .node_string
+                            .as_ref()
+                            .unwrap()
+                            .split("::")
+                            .last()
+                            .unwrap()
+                    );
+                    accum += &child.1.recurse_serialize_to_reexports(indent + 1);
+                    accum += &format!("{}}}\n", indent_str);
+                }
+            };
+        });
+
+        accum
+    }
+}
